Rewrite communication structure, proceed command from can bus

+++ DrivingSystem/BLDC_program/pid_control.h
@@ -11,7 +11,7 @@ typedef struct{
   double decayRatio;
   
   /* Computing Parameter */
-  uint16_t dt;
+  double dt;
   int16_t upperBound;
   int16_t lowerBound;
